requirements.txt
tabulate==0.9.0
pyserial==3.5